WIP: core/mount: Don't RemoveAll if unmount failed
This is to further debug issue #10704.

We probably want to change this to use os.Remove() that will only remove
empty dirs at least, but let's confirm this is the issue and we can see
what the right way to do it is later. It will be great if we can avoid
leaking mounts in /tmp/ in case of errors.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>